+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 * Add "Add argument to predicate/non-terminal" refactoring support
 * Add "Reorder predicate/non-terminal arguments" refactoring support
 * Add "Remove argument from predicate/non-terminal" refactoring support
+* Add "Add parameter to object/category" refactoring support
+* Add "Reorder object/category parameters" refactoring support
+* Add "Remove parameter from object/category" refactoring support
 * Add "Extract protocol" refactoring support
 * Add "Extract to new Logtalk entity" refactoring support
 * Add "Extract to new Logtalk file" refactoring support